requirements.txt
easysnmp==0.2.6
PyYAML==6.0.1
pysmi==0.3.4